+++ base/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Require [`@typescript-eslint/eslint-plugin`](https://www.npmjs.com/package/@typescript-eslint/eslint-plugin) `^3.0.0` as a peer dependency
   - Some rules have been turned on. See [the change log of the library](https://github.com/typescript-eslint/typescript-eslint/releases/tag/v3.0.0) for details.
 - Require [`@typescript-eslint/parser`](https://www.npmjs.com/package/@typescript-eslint/parser) `^3.0.0` as a peer dependency
+- Turn on [`@typescript-eslint/naming-convention`](https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/naming-convention.md)
 
 ## v0.1.0